+++ tests/fixtures/svelte5-runes/NoInterface/output.d.ts
@@ -1,0 +1,14 @@
+import { SvelteComponentTyped } from "svelte";
+
+export type Svelte5RunesNoInterfaceProps = {
+  /**
+   * @default undefined
+   */
+  foo: string;
+};
+
+export default class Svelte5RunesNoInterface extends SvelteComponentTyped<
+  Svelte5RunesNoInterfaceProps,
+  Record<string, any>,
+  Record<string, never>
+> {}